requirements.txt
anthropic>=0.40.0
pymupdf>=1.24.0
python-dotenv>=1.0.0

# Google Drive monitor
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
openpyxl>=3.1.0

# Siigo API
requests>=2.31.0

# Dashboard web
streamlit>=1.35.0
pandas>=2.0.0
